fixed bug: amount in groupview now shows decimals

+++ app/src/main/java/com/zj/android/expensetracker/DatabaseAccessor.java
@@ -76,7 +76,7 @@ public class DatabaseAccessor {
 
         try {
             cursor.moveToFirst();
-            return cursor.getInt("TotalAmount");
+            return (float) cursor.getDouble("TotalAmount");
         } finally {
             cursor.close();
         }
